LayoutTests/fast/forms/select-namedItem-expected.txt

Test for HTMLSelectElement.namedItem()

FOUND ITEM: 2
FOUND ITEM: 1
FOUND ITEM: 1
FOUND ITEM: 1
ERROR: Null value
